+++ sdk/communication/azure-communication-chat/env-prod.properties
@@ -6,4 +6,4 @@ platform.ui.version="0.0.0.0"
 trouter.hostname="go.trouter.skype.com"
 trouter.registration.hostname.and.base.path="edge.skype.com/registrar/prod"
 trouter.application.id="AcsAndroid"
-trouter.template.key="AcsAndroid_Chat_1.1"
+trouter.template.key="AcsAndroid_Chat_1.2"
