+++ sdk/communication/azure-communication-chat/env-test.properties
@@ -6,4 +6,4 @@ platform.ui.version="0.0.0.0"
 trouter.hostname="go.trouter-int.skype.net"
 trouter.registration.hostname.and.base.path="edge.skype.net/registrar/testenv"
 trouter.application.id="AcsAndroid_test"
-trouter.template.key="AcsAndroid_Chat_test_1.1"
+trouter.template.key="AcsAndroid_Chat_test_1.2"
