+++ sdk/communication/azure-communication-chat/src/main/java/com/azure/android/communication/chat/ChatAsyncClient.java
@@ -12,9 +12,9 @@ import com.azure.android.communication.chat.models.ChatThreadItem;
 import com.azure.android.communication.chat.models.CreateChatThreadOptions;
 import com.azure.android.communication.chat.models.CreateChatThreadResult;
 import com.azure.android.communication.chat.models.ListChatThreadsOptions;
-import com.azure.android.communication.chat.signaling.RealTimeNotificationCallback;
-import com.azure.android.communication.chat.signaling.SignalingClient;
-import com.azure.android.communication.chat.signaling.properties.ChatEventId;
+import com.azure.android.communication.chat.models.signaling.RealTimeNotificationCallback;
+import com.azure.android.communication.chat.implementation.signaling.SignalingClient;
+import com.azure.android.communication.chat.models.signaling.ChatEventId;
 import com.azure.android.core.logging.ClientLogger;
 import com.azure.android.core.rest.Page;
 import com.azure.android.core.rest.PagedResponse;
