+++ sdk/communication/azure-communication-chat/src/main/java/com/azure/android/communication/chat/implementation/signaling/ConnectionState.java
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.android.communication.chat.signaling;
+package com.azure.android.communication.chat.implementation.signaling;
 
 /**
  * Connection states
