+++ sdk/communication/azure-communication-chat/src/main/java/com/azure/android/communication/chat/implementation/signaling/SignalingClient.java
@@ -1,9 +1,10 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.android.communication.chat.signaling;
+package com.azure.android.communication.chat.implementation.signaling;
 
-import com.azure.android.communication.chat.signaling.properties.ChatEventId;
+import com.azure.android.communication.chat.models.signaling.ChatEventId;
+import com.azure.android.communication.chat.models.signaling.RealTimeNotificationCallback;
 
 /**
  * The signaling client interface
