+++ sdk/communication/azure-communication-chat/src/main/java/com/azure/android/communication/chat/models/signaling/ChatEventId.java
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.android.communication.chat.signaling.properties;
+package com.azure.android.communication.chat.models.signaling;
 
 /**
  * Defines values for chat event.
@@ -16,6 +16,5 @@ public enum ChatEventId {
     chatThreadDeleted,
     chatThreadPropertiesUpdated,
     participantsAdded,
-    participantsRemoved,
-    connectionChanged;
+    participantsRemoved
 }
