+++ sdk/communication/azure-communication-chat/src/main/java/com/azure/android/communication/chat/models/signaling/ChatMessageEditedEvent.java
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.android.communication.chat.signaling.chatevents;
+package com.azure.android.communication.chat.models.signaling;
 
 /**
  * Event for a edited chat message.
@@ -40,7 +40,6 @@ public class ChatMessageEditedEvent extends ChatUserEvent {
      * `1593117207131`
      */
     private String version;
-
 
     /**
      * Gets The timestamp when the message was edited. The timestamp is in ISO8601 format:
