+++ sdk/communication/azure-communication-chat/src/main/java/com/azure/android/communication/chat/models/signaling/ChatMessageReceivedEvent.java
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.android.communication.chat.signaling.chatevents;
+package com.azure.android.communication.chat.models.signaling;
 
 /**
  * Event for a received chat message.
