+++ sdk/communication/azure-communication-chat/src/main/java/com/azure/android/communication/chat/models/signaling/ChatThreadEvent.java
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.android.communication.chat.signaling.chatevents;
+package com.azure.android.communication.chat.models.signaling;
 
 /**
  * Chat thread event
@@ -14,7 +14,6 @@ public abstract class ChatThreadEvent extends BaseEvent {
      * `1593117207131`
      */
     private String version;
-
 
     /**
      * Sets new Version of the thread. This version is an epoch time in a numeric unsigned Int64 format:
