+++ sdk/communication/azure-communication-chat/src/main/java/com/azure/android/communication/chat/models/signaling/ChatUserEvent.java
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.android.communication.chat.signaling.chatevents;
+package com.azure.android.communication.chat.models.signaling;
 
 import com.azure.android.communication.common.CommunicationIdentifier;
 
@@ -19,7 +19,6 @@ public abstract class ChatUserEvent extends BaseEvent {
      * The Id of the event recipient.
      */
     private CommunicationIdentifier recipient;
-
 
     /**
      * Sets new The Id of the event recipient..
