+++ sdk/communication/azure-communication-chat/src/main/java/com/azure/android/communication/chat/models/signaling/ParticipantsRemovedEvent.java
@@ -1,9 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.android.communication.chat.signaling.chatevents;
+package com.azure.android.communication.chat.models.signaling;
 
-import com.azure.android.communication.chat.signaling.properties.ChatParticipant;
+import com.azure.android.communication.chat.models.ChatParticipant;
 
 import java.util.List;
 
@@ -27,7 +27,6 @@ public class ParticipantsRemovedEvent extends ChatThreadEvent {
      * The information of the user that removed the chat participants.
      */
     private ChatParticipant removedBy;
-
 
     /**
      * Gets The participants removed from the thread..
