+++ sdk/communication/azure-communication-chat/src/main/java/com/azure/android/communication/chat/models/signaling/RealTimeNotificationCallback.java
@@ -1,13 +1,10 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.android.communication.chat.signaling;
-
-import com.azure.android.communication.chat.signaling.chatevents.BaseEvent;
-
+package com.azure.android.communication.chat.models.signaling;
 
 /**
- *  Funcitonal interface of real time notification callback
+ *  Functional interface of real time notification callback
  */
 @FunctionalInterface
 public interface RealTimeNotificationCallback {
