+++ sdk/communication/azure-communication-chat/src/main/java/com/azure/android/communication/chat/models/signaling/package-info.java
@@ -3,5 +3,5 @@
 // license information.
 //
 
-/** Package containing the classes for Signaling properties. */
-package com.azure.android.communication.chat.signaling.properties;
+/** Package containing the classes for Chat events classes */
+package com.azure.android.communication.chat.models.signaling;
