--- sdk/communication/azure-communication-chat/src/main/java/com/azure/android/communication/chat/signaling/package-info.java
@@ -1,7 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License. See License.txt in the project root for
-// license information.
-//
-
-/** Package containing the classes for Signaling. Real time notification service. */
-package com.azure.android.communication.chat.signaling;
